refactor: extract logic into modules
